src/requirements.txt
# Add function packages requirements here
google-cloud-bigquery